fix `parse_any_delimiter` lifetimes
Makes `parse_any_delimiter` return a `ParseBuffer` with the correct lifetime.
It previously used the lifetime of the reference to self but the ParseBuffer is valid as long as it's cursor is and the cursor is valid for as long as the cursor it's derived from is, not some arbitrary reference to it.